chore: #91 move txn_id to separate file
